Move the embed transformer into a standalone function
