+++ content/webapp/services/prismic/types/body.ts
@@ -83,7 +83,7 @@ export type Embed = Slice<
   'embed',
   {
     embed: EmbedField;
-    caption: KeyTextField;
+    caption: RichTextField;
   }
 >;
